refactor: 패키지 이동과 Contract 내부로 View 인터페이스 이동
