+++ app/src/androidTest/java/woowacourse/movie/view/reservation/ReservationActivityLandscapeTest.kt
@@ -17,6 +17,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.movie.R
+import woowacourse.movie.feature.reservation.ReservationActivity
 import woowacourse.movie.view.model.ResourceMapper
 import woowacourse.movie.view.model.ScreeningData
 import java.time.LocalDate
