+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,16 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".view.reservation.ReservationActivity"
+            android:name=".feature.reservation.ReservationActivity"
             android:exported="false" />
         <activity
-            android:name=".view.ticket.TicketActivity"
+            android:name=".feature.ticket.TicketActivity"
             android:exported="false" />
         <activity
-            android:name=".view.selectSeat.SelectSeatActivity"
+            android:name=".feature.selectseat.SelectSeatActivity"
             android:exported="false" />
         <activity
-            android:name=".view.main.MainActivity"
+            android:name=".feature.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
