+++ app/src/main/java/woowacourse/movie/feature/main/MainPresenter.kt
@@ -1,17 +1,16 @@
-package woowacourse.movie.presenter
+package woowacourse.movie.feature.main
 
 import woowacourse.movie.model.movie.Advertisement
 import woowacourse.movie.model.movie.screening.Screening
-import woowacourse.movie.view.main.MainView
 import woowacourse.movie.view.model.ImageResource
 import woowacourse.movie.view.model.ScreeningData
 
 class MainPresenter(
-    private val view: MainView,
+    private val view: MainContract.View,
 ) {
     fun initMainUI() {
-        val screenings: List<Screening> = Screening.getDefaultScreenings()
-        val ads: List<Advertisement> = Advertisement.getDefaultAds()
+        val screenings: List<Screening> = Screening.Companion.getDefaultScreenings()
+        val ads: List<Advertisement> = Advertisement.Companion.getDefaultAds()
         view.initMovieListUI(screenings, ads)
     }
