+++ app/src/main/java/woowacourse/movie/feature/main/MovieAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.main
+package woowacourse.movie.feature.main
 
 import android.view.LayoutInflater
 import android.view.View
