+++ app/src/main/java/woowacourse/movie/feature/reservation/ReservationContract.kt
@@ -1,0 +1,33 @@
+package woowacourse.movie.feature.reservation
+
+import woowacourse.movie.model.movie.screening.Screening
+import woowacourse.movie.model.ticket.TicketCount
+import woowacourse.movie.view.model.ScreeningData
+import woowacourse.movie.view.model.TicketData
+import java.time.LocalDate
+
+interface ReservationContract {
+    interface View {
+        fun getScreeningData(): ScreeningData
+
+        fun initScreeningInfoUI(screeningData: ScreeningData)
+
+        fun setDateSelectUi(screening: Screening)
+
+        fun setTimeSelectUi(
+            selectedDate: LocalDate,
+            screening: Screening,
+            position: Int,
+        )
+
+        fun setTicketCounterUi(ticketCount: TicketCount)
+
+        fun initTicketPlusBtnUi()
+
+        fun initTicketMinusBtnUi()
+
+        fun printError(message: String)
+
+        fun navigateToSelectSeatUI(ticketData: TicketData)
+    }
+}
