+++ app/src/main/java/woowacourse/movie/feature/ticket/TicketContract.kt
@@ -1,0 +1,12 @@
+package woowacourse.movie.feature.ticket
+
+import woowacourse.movie.model.ticket.Ticket
+import woowacourse.movie.view.model.TicketData
+
+interface TicketContract {
+    interface TicketView {
+        fun getTicketData(): TicketData
+
+        fun initTicketUI(ticket: Ticket)
+    }
+}
