+++ app/src/main/java/woowacourse/movie/feature/ticket/TicketPresenter.kt
@@ -1,15 +1,14 @@
-package woowacourse.movie.presenter
+package woowacourse.movie.feature.ticket
 
 import woowacourse.movie.model.ticket.Ticket
 import woowacourse.movie.model.ticket.TicketCount
 import woowacourse.movie.model.ticket.TicketPrice
 import woowacourse.movie.model.ticket.getOrDefault
 import woowacourse.movie.view.model.TicketData
-import woowacourse.movie.view.ticket.TicketView
 import java.time.LocalDateTime
 
 class TicketPresenter(
-    private val view: TicketView,
+    private val view: TicketContract.TicketView,
 ) {
     private val ticketData: TicketData by lazy {
         view.getTicketData()
@@ -20,7 +19,7 @@ class TicketPresenter(
         val showtime: LocalDateTime = ticketData.showtime
         Ticket(
             screening = screening,
-            ticketCount = TicketCount.create(ticketCount).getOrDefault(),
+            ticketCount = TicketCount.Companion.create(ticketCount).getOrDefault(),
             showtime = showtime,
         )
     }
